remove NpyCharArray
as we do not really have a use case for it (currently). Add
`dataType` and `writeElementTo(i, buffer)` to the array types.
not sure if we should leeak this into the API... maybe we should
move the array types back to the main npy package so that these
methods can be package private...